Release 3.0.0
This release corrects an issue with the license listed in the
package.json. There is no changes to code.

The license listed in the package.json is BSD-2-Clause.
But the license in the LICENSE.md file and the license
attached to the original project at https://github.com/facebook/fbjs
is BSD-3-clause.

This release corrects the license in the package.json to reflect the
original.